test: move import_helper out
